+++ CHANGELOG.md
@@ -6,8 +6,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.4.0] - 2018-05-24
 ### Changed
-- Change the mini cart components to use the store components.
+- The mini cart components to use the store components.
+- PropTypes to reuse the mini cart's one.
 
 ## [0.3.2] - 2018-05-19
 ### Fixed